lovable/backend/requirements.txt
# Lovable Backend Requirements

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic[email]>=2.5.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.0  # PostgreSQL sync driver
sqlite3  # Built-in, but listed for clarity

# Authentication & Security
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
cryptography>=41.0.0

# HTTP & Networking
httpx>=0.25.0
requests>=2.31.0
urllib3>=2.0.0

# Redis (optional, for caching/sessions)
redis>=5.0.0
hiredis>=2.2.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Validation & Serialization
email-validator>=2.1.0
phonenumbers>=8.13.0

# Date/Time
python-dateutil>=2.8.0
pytz>=2023.3

# Utilities
typing-extensions>=4.8.0
click>=8.1.0
rich>=13.6.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing
faker>=20.1.0  # For test data

# Code Quality
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocs-swagger-ui-tag>=0.6.0

# Monitoring & Logging
structlog>=23.2.0
prometheus-client>=0.19.0
sentry-sdk[fastapi]>=1.38.0

# File Processing
python-magic>=0.4.27
pillow>=10.1.0
pandas>=2.1.0  # For data processing
openpyxl>=3.1.0  # Excel support
python-docx>=1.1.0  # Word document support
PyPDF2>=3.0.0  # PDF processing

# Task Queue (optional)
celery>=5.3.0
redis>=5.0.0  # Backend for Celery

# WebSocket Support
websockets>=12.0

# Template Engine (optional)
jinja2>=3.1.0

# Internationalization (optional)
babel>=2.13.0

# Environment-specific
# Production
gunicorn>=21.2.0
gevent>=23.9.0

# Development
watchdog>=3.0.0
